+++ src/main/java/de/samply/frontend/dto/configuration/ProjectConfiguration.java
@@ -1,0 +1,11 @@
+package de.samply.frontend.dto.configuration;
+
+import lombok.Data;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Data
+public class ProjectConfiguration {
+    private Map<String, String> fieldValues = new HashMap<>();
+}