+++ src/main/java/de/samply/frontend/dto/configuration/ProjectConfigurations.java
@@ -1,0 +1,30 @@
+package de.samply.frontend.dto.configuration;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import de.samply.frontend.dto.Project;
+import lombok.Data;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Data
+public class ProjectConfigurations {
+
+    private Map<String, ProjectConfiguration> config = new HashMap<>();
+    @JsonIgnore
+    private Map<String, Project> configurationNameProjectMap = new HashMap<>();
+
+    @JsonIgnore
+    public void initConfigurationNameProjectMap() {
+        config.forEach((configurationName, projectConfiguration) -> configurationNameProjectMap.put(configurationName, convert(projectConfiguration)));
+    }
+
+    @JsonIgnore
+    private static Project convert(ProjectConfiguration projectConfiguration) {
+        Project result = new Project();
+        result.setFieldsFromMap(projectConfiguration.getFieldValues());
+        return result;
+    }
+
+
+}