+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2024-02-26]
+## [0.0.1 - 2024-02-29]
 ### Added
 - First version of the project
 - Spring Application
@@ -109,3 +109,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Exists Script
 - Explorer Redirect URI
 - User State in Frontend DTO
+- Fetch project configurations